Avoid mild precision issue in hex grid offsets
Previously we were offseting by 3/8 on the U axis, but this is only correct for regular hexagons. We now express this
more generally in terms of edge length and projection. FIXME

Also some clean up:
- One implementation for both types of hex grids
- Don't offset by more than a grid cell. It's unnecessary and means we need to do so everywhere else too.
